+++ .changeset/jovial-jaybirds-jaunter.md
@@ -1,0 +1,3 @@
+---
+'@finos/legend-art': patch
+---
