+++ .changeset/meek-mockingbirds-minister.md
@@ -1,0 +1,3 @@
+---
+'@finos/legend-graph': patch
+---
