+++ .changeset/plentiful-pigeons-preen.md
@@ -1,0 +1,3 @@
+---
+'@finos/legend-application-studio': patch
+---
