+++ .changeset/shiny-rabbits-train.md
@@ -1,3 +1,5 @@
 ---
 '@finos/legend-query-builder': patch
 ---
+
+Add show sql option allowing users to view generated sql for a query.
