+++ packages/legend-art/style/base/_panel.scss
@@ -138,6 +138,7 @@
 
     &--dark {
       color: var(--color-light-grey-50);
+      background: var(--color-dark-grey-50);
     }
   }
