+++ packages/legend-graph/src/graph/helpers/DomainHelper.ts
@@ -69,6 +69,26 @@ import { Multiplicity } from '../metamodel/pure/packageableElements/domain/Multi
 import type { AnnotatedElement } from '../metamodel/pure/packageableElements/domain/AnnotatedElement.js';
 import type { ConcreteFunctionDefinition } from '../metamodel/pure/packageableElements/domain/ConcreteFunctionDefinition.js';
 import { extractDependencyGACoordinateFromRootPackageName } from '../DependencyManager.js';
+import {
+  type RelationalDataType,
+  VarChar,
+  Char,
+  VarBinary,
+  Binary,
+  Bit,
+  Numeric,
+  Decimal,
+  Double,
+  Float,
+  Real,
+  Integer,
+  SmallInt,
+  TinyInt,
+  Timestamp,
+  Other,
+  SemiStructured,
+  Json,
+} from '../../STO_Relational_Exports.js';
 
 export const addElementToPackage = (
   parent: Package,
@@ -539,6 +559,49 @@ export const isSubType = (type1: Type, type2: Type): boolean => {
     );
   }
   return false;
+};
+
+export const stringifyDataType = (type: RelationalDataType): string => {
+  if (type instanceof VarChar) {
+    return `VARCHAR(${type.size})`;
+  } else if (type instanceof Char) {
+    return `CHAR(${type.size})`;
+  } else if (type instanceof VarBinary) {
+    return `VARBINARY(${type.size})`;
+  } else if (type instanceof Binary) {
+    return `BINARY(${type.size})`;
+  } else if (type instanceof Bit) {
+    return `BIT`;
+  } else if (type instanceof Numeric) {
+    return `NUMERIC(${type.precision},${type.scale})`;
+  } else if (type instanceof Decimal) {
+    return `DECIMAL(${type.precision},${type.scale})`;
+  } else if (type instanceof Double) {
+    return `DOUBLE`;
+  } else if (type instanceof Float) {
+    return `FLOAT`;
+  } else if (type instanceof Real) {
+    return `REAL`;
+  } else if (type instanceof Integer) {
+    return `INTEGER`;
+  } else if (type instanceof BigInt) {
+    return `BIGINT`;
+  } else if (type instanceof SmallInt) {
+    return `SMALLINT`;
+  } else if (type instanceof TinyInt) {
+    return `TINYINT`;
+  } else if (type instanceof Date) {
+    return `DATE`;
+  } else if (type instanceof Timestamp) {
+    return `TIMESTAMP`;
+  } else if (type instanceof Other) {
+    return `OTHER`;
+  } else if (type instanceof SemiStructured) {
+    return 'SEMI-STRUCTURED';
+  } else if (type instanceof Json) {
+    return 'JSON';
+  }
+  return '(UNKNOWN)';
 };
 
 /**
