+++ packages/legend-query-builder/src/components/QueryBuilderResultPanel.tsx
@@ -91,7 +91,6 @@ import {
   CodeEditor,
 } from '@finos/legend-lego/code-editor';
 import { ExecutionPlanViewer } from './execution-plan/ExecutionPlanViewer.js';
-import { SqlQueryViewer } from './sql-query/SqlQueryViewer.js';
 
 const QueryBuilderGridResultContextMenu = observer(
   forwardRef<
@@ -723,7 +722,6 @@ export const QueryBuilderResultPanel = observer(
         <ExecutionPlanViewer
           executionPlanState={resultState.executionPlanState}
         />
-        <SqlQueryViewer resultState={resultState} />
       </div>
     );
   },
