+++ packages/legend-query-builder/src/components/execution-plan/SQLExecutionNodeViewer.tsx
@@ -15,13 +15,19 @@
  */
 
 import { observer } from 'mobx-react-lite';
-import type { ExecutionPlanState } from '../../stores/execution-plan/ExecutionPlanState.js';
-import { format as formatSQL } from 'sql-formatter';
-import type { SQLResultColumn } from '@finos/legend-graph';
 import {
-  CODE_EDITOR_LANGUAGE,
-  CodeEditor,
-} from '@finos/legend-lego/code-editor';
+  type ExecutionPlanState,
+  EXECUTION_PLAN_VIEW_MODE,
+} from '../../stores/execution-plan/ExecutionPlanState.js';
+import { format as formatSQL } from 'sql-formatter';
+import { type SQLResultColumn, stringifyDataType } from '@finos/legend-graph';
+import {
+  PanelListItem,
+  CopyIcon,
+  PanelDivider,
+  Button,
+  PanelContent,
+} from '@finos/legend-art';
 
 /**
  * TODO: Create a new `AbstractPlugin` for this, called `ExecutionPlanViewerPlugin`
@@ -35,14 +41,95 @@ export const SQLExecutionNodeViewer: React.FC<{
   resultColumns: SQLResultColumn[];
   executionPlanState: ExecutionPlanState;
 }> = observer((props) => {
-  const { query } = props;
+  const { query, resultColumns, executionPlanState } = props;
 
+  const applicationStore = executionPlanState.applicationStore;
+  const copyExpression = (value: string): void => {
+    applicationStore.clipboardService
+      .copyTextToClipboard(value)
+      .then(() =>
+        applicationStore.notificationService.notifySuccess(
+          'SQL Query copied',
+          undefined,
+          2500,
+        ),
+      )
+      .catch(applicationStore.alertUnhandledError);
+  };
   return (
-    <CodeEditor
-      inputValue={formatSQL(query)}
-      isReadOnly={true}
-      language={CODE_EDITOR_LANGUAGE.SQL}
-      hideMinimap={true}
-    />
+    <PanelContent
+      darkMode={
+        !applicationStore.layoutService.TEMPORARY__isLightColorThemeEnabled
+      }
+    >
+      <div className="query-builder__sql__container">
+        <PanelDivider />
+        <div key={query}>
+          <div className="query-builder__sql__container__item__label">
+            <PanelListItem>
+              SQL
+              <div>
+                <button
+                  onClick={() => {
+                    copyExpression(query);
+                  }}
+                  title="Copy SQL Expression"
+                  className="query-builder__sql__container__icon"
+                >
+                  <CopyIcon />
+                </button>
+              </div>
+            </PanelListItem>
+          </div>
+          <PanelListItem className="query-builder__sql__container__item">
+            <pre>{formatSQL(query)} </pre>
+          </PanelListItem>
+          <PanelDivider />
+        </div>
+      </div>
+      <div className="query-builder__sql__container">
+        <div>
+          <PanelListItem className="query-builder__sql__container__item__label">
+            Result Columns
+          </PanelListItem>
+          <PanelDivider />
+          <table className="table query-builder__sql__container__table">
+            <thead>
+              <tr>
+                <th className="table__cell--left">Label</th>
+                <th className="table__cell--left">Data Type</th>
+              </tr>
+            </thead>
+            <tbody>
+              {resultColumns.map((column) => (
+                <tr key={column.label}>
+                  <td className="table__cell--left">
+                    {column.label.replaceAll(`"`, '')}
+                  </td>
+
+                  {column.dataType && (
+                    <td className="table__cell--left">
+                      {stringifyDataType(column.dataType)}
+                    </td>
+                  )}
+                </tr>
+              ))}
+            </tbody>
+          </table>
+        </div>
+      </div>
+      <PanelDivider />
+
+      <div className="query-builder__sql__container">
+        <Button
+          className="btn--dark execution-node-viewer__unsupported-view__to-text-mode__btn"
+          onClick={(): void =>
+            executionPlanState.setViewMode(EXECUTION_PLAN_VIEW_MODE.JSON)
+          }
+          text="View JSON"
+        />
+      </div>
+      <PanelDivider />
+    </PanelContent>
   );
 });
